refactor: simplify erroboudnary:
