+++ src/app/App.tsx
@@ -1,13 +1,10 @@
 import { MainPage } from '../pages/MainPage/ui/MainPage';
-import { ErrorBoundary } from './errorBoundary/ErrorBoundary';
 import '../shared/styles';
 
 function App() {
   return (
     <>
-      <ErrorBoundary>
-        <MainPage></MainPage>
-      </ErrorBoundary>
+      <MainPage />
     </>
   );
 }
