--- src/features/ErrorText/index.ts
@@ -1,1 +1,0 @@
-export { ErrorText } from './ui/ErrorText';
